# use ShapeType ordinals when deserializing shape trees by type