StreamInput#readEnum allocates objects when retrieving class constants
from the ordinal value it serialized in writeEnum. This change
by-passes this call to reduce the amount of objects allocated  at read-time.


allocation/performance results of tiling the Switzerland polygon
at precision 9.

before PR:

```
GeoTileGridBenchmark.measureGeoTile_GeometryTree                               avgt    8  35821.991 ± 1722.838   ns/op
GeoTileGridBenchmark.measureGeoTile_GeometryTree:·gc.alloc.rate                avgt    8    253.492 ±   12.552  MB/sec
GeoTileGridBenchmark.measureGeoTile_GeometryTree:·gc.alloc.rate.norm           avgt    8  14304.016 ±    0.004    B/op
GeoTileGridBenchmark.measureGeoTile_GeometryTree:·gc.churn.G1_Eden_Space       avgt    8    248.932 ±  176.201  MB/sec
GeoTileGridBenchmark.measureGeoTile_GeometryTree:·gc.churn.G1_Eden_Space.norm  avgt    8  14015.221 ± 9644.184    B/op
GeoTileGridBenchmark.measureGeoTile_GeometryTree:·gc.churn.G1_Old_Gen          avgt    8      0.019 ±    0.052  MB/sec
GeoTileGridBenchmark.measureGeoTile_GeometryTree:·gc.churn.G1_Old_Gen.norm     avgt    8      1.063 ±    2.932    B/op
GeoTileGridBenchmark.measureGeoTile_GeometryTree:·gc.count                     avgt    8     10.000             counts
GeoTileGridBenchmark.measureGeoTile_GeometryTree:·gc.time                      avgt    8     25.000                 ms
```

With PR:

```
GeoTileGridBenchmark.measureGeoTile_GeometryTree                               avgt    8  35117.419 ± 1570.024   ns/op
GeoTileGridBenchmark.measureGeoTile_GeometryTree:·gc.alloc.rate                avgt    8    172.580 ±    7.578  MB/sec
GeoTileGridBenchmark.measureGeoTile_GeometryTree:·gc.alloc.rate.norm           avgt    8   9544.015 ±    0.003    B/op
GeoTileGridBenchmark.measureGeoTile_GeometryTree:·gc.churn.G1_Eden_Space       avgt    8    149.195 ±  176.064  MB/sec
GeoTileGridBenchmark.measureGeoTile_GeometryTree:·gc.churn.G1_Eden_Space.norm  avgt    8   8218.159 ± 9708.295    B/op
GeoTileGridBenchmark.measureGeoTile_GeometryTree:·gc.churn.G1_Old_Gen          avgt    8      0.013 ±    0.061  MB/sec
GeoTileGridBenchmark.measureGeoTile_GeometryTree:·gc.churn.G1_Old_Gen.norm     avgt    8      0.715 ±    3.311    B/op
GeoTileGridBenchmark.measureGeoTile_GeometryTree:·gc.count                     avgt    8      6.000             counts
GeoTileGridBenchmark.measureGeoTile_GeometryTree:·gc.time                      avgt    8     15.000                 ms
```

This benchmark demonstrates a 2% performance improvement with a 40% reduction in gc-counts / allocation-bytes